Log extlink action when appropriate
A small number of Zero operators whitelist only certain language wikis.
This action logs when a user is not shown an exit interstitial despite
being potentially zero-rated based on IP address because the specific wiki
the user was on was in a non-zero-rated language.

An example of this is when a user is on a network owned by a provider that
zero-rates only English and French Wikipedias, and is browsing German
Wikipedia and clicks an external link.

Bug: T123748
Change-Id: Ia09b03c9e435303240e1ffe5a6e124827bcda582