Fix memory leaks (for real!), part 1.
My ongoing search has finally yielded some tangible nuggets.
Exhibit A: Adding a TextWatcher to a TextView, but never removing it! (The
TextWatcher holds a reference to the fragment, thereby leaking it)

Also, now explicitly removing the onItemClick and onItemLongClick
listeners from the ListView, since this seems to lead to a leak in
API <20.

Change-Id: I497ebd37f3059b420622416a66ee5b87a15a2d0c