Hygiene: Rename Restbase server classes to match MW API server classes
While writing the Wiktionary definition endpoint I had the idea that a
definition for our purposes was notionally separate from a page, and so
I renamed these.  But this doesn't reflect the endpoint structure (which
still contains /page/) and ruins the correspondence with the MW API server
classes.  So it's better to change them back.

Change-Id: Icc848f4fddc56b400efdf7214bd058d574836bda