Add X-Analytics request headers for previews and pageviews
preview=1 is set whenever the summary endpoints are used
pageview=1 is set whenever the lead section requests are used

This is done for both MW API and RESTBase endpoints.
For the MW API we need to add it to the lead section request since the
Analytics Refinery uses the legacy lead section requests.
(We don't want to count both the lead section request and the remaining
section request as separate pageviews.)
For consistency sake I did the same for the RESTBase case, even though
there it's not strictly required.

Bug: T128612
Change-Id: I497691fed1ab30d62a7f3359503bc099e0be0f02