Upgrade database versions sequentially
Database upgrades are exceptional. Upgrading differently based on
originating version is a small performance gain for a large liability
and test matrix. This patch changes upgrades to always follow the same
path with each version upgraded sequentially until the current db
version is met. This means that if an upgrade on the oldest version of
the app works, all the upgrades in between should too. Additionally,
table upgrade logic was pushed from the abstract super into the subclass
implementation to keep responsibilities separate.

Upgrades between each version work prior to this patch so testing was
performed from each of the following versions directly to this patch:

  • 2.0 (2014-06-25)
  • 2.0 (2015-03-23)
  • 2.0.102 (2015-05-14)
  • 2.0.107 (2015-08-03)
  • 2.0.110 (2015-08-31)
  • 2.1.136 (2015-12-09)
  • 2.1.140 (2016-02-04)
  • 2.1.141 (2016-02-10)
  • 2.1.142 (2016-03-07)

Change-Id: Id540478abf4d110840d92e43a13bd7fc0441caad