Upgrade to Support 23.2.1, without functional changes.
The only necessary workarounds I had to make have to do with the system
bar, and the inability to set its flags via styles. It looks like the
"windowTranslucentStatus" and "windowFullScreen" flags no longer work when
set via an inherited theme, so I'm setting them in the code.

Also, the "save page" icon in the toolbar wasn't getting the proper tint
(possibly a bug with tinting state drawables?), so I changed the bookmark
drawable itself to be explicitly white, since it's not used anywhere else.

Also, the new version seems to have broken the
getResources().getDisplayMetrics() function, where it fails to update the
metrics upon screen rotation. I made a workaround for this, as well.
https://code.google.com/p/android/issues/detail?id=201928

Bug: T128126
Change-Id: I5d231a113e36d8acec973fc4a9f0ece2eaa58d7e