Improve link preview thumbnail gallery alignment.
The new version of the Support library allows RecyclerViews to have
"wrap_content" dimensions. In our link preview window, this means that we
can now center-align the thumbnail gallery when there are fewer images
than the width of the dialog.

I'm therefore removing the minimum number of images shown in the thumbnail
gallery.

To test: go to the [[Earth]] article, and click on the [[Greek]] link in
the first sentence.  Thumbnail galleries with three or more images should
not be affected.

Change-Id: I89b2e55eedc5e4b5380d8f159ee2abf9fd9b151e