Prevent WebView loading spinner from persisting inappropriately
In certain cases (namely, for very short pages), the call to
bridge.sendMessage at the end of the displayLeadSection listener callback
could trigger the execution of a pending JavaScript callback related to
the remaining sections, with the practical effect that the loading spinner
would be removed before being added.  This changes the order of execution
so that doesn't happen.

Bug: T123940
Change-Id: I404d2d1580021de45e3590bf0d15a44ee5811acc