Fix launching URLs in external browser for API 23.
In API 23, once an app is set as the "default" for handing a certain Uri,
then the app will be the only one returned by the queryIntentActivities()
function, which breaks the expected behavior of this function in previous
APIs.

We fix this issue by explicitly setting a temporary Uri when querying
intent activities, to ensure that we get the default system browser
activities.

Bug: T132192
Change-Id: Ic7f2deccf6f327bf69bd240d9ceb8356cdd959c4