Replace erroneously removed cite styles
Generating these requires the 'cite' role in mw-vagrant, which I didn't
have before (leading to an error generating the CSS).

Thus, I removed the reference from the CSS generation script.

This puts back the needed styles.  I'll update the mobileapp role in a
follow-up patch to require the cite role.

Bug: T132253
Change-Id: I7a9c8ef804301bc81932886a2aec74fe96ba3f84