Add filename to Reading List page disk table
Add a filename column to the Reading List page disk database table. This
column is necessary to remember where files were stored.

The previous database schema has not been published through official
channels yet so the effort to upgrade from the previous to the new
has not been made. If you have installed a dev build with the current
schema, this change will cause a crash or undefined behavior. Please
clear your data.

Change-Id: I0e2c72e1b1d632b6be75bbd870d1b03ebcb690a7