Fix bug when entering Edit activity without highlighting text.
Right now, tapping the "edit" pencil launches the Edit activity and shows
an error and "retry" button overlaid onto the wikitext. This is due to an
exception thrown from the auto-highlighting code, trying to highlight a
null string.

The issue is that the isGraphic() function itself doesn't check for null.

Change-Id: I8bcce5ee8b9a7a9e9a065ac91733b06b5196feac