Improve clickability of links in Wiktionary dialog.
The issue was that the Wiktionary dialog expects all links to start with
"/wiki/", whereas certain links start with "./". I expanded the link
parsing to support this case, as well.

In addition, the Wiktionary dialog couldn't handle links that have a
fragment in them. We'll simply ignore fragments for now.

Bug: T132397
Change-Id: I13daccdc26802f435d07b416a46154c55933dcf0