Vector drawables, part 2.
Replace most other raster icons with vector equivalents, performing a
little cleanup along the way.

The remaining raster icons are either "custom" graphics specific to WMF,
or graphics that need to be flipped for RTL. Android vector drawables are
indeed flippable for RTL, but only for API 21+, so we'll need to keep
these for now.

Bug: T128046
Change-Id: Ie510342f0590f19e034ea5410ca58a4757ca3394