Fix duplicate Search icon for RTL in API <23
There are two places in our code where we use "drawableStart" to insert a
drawable at the start of a TextView. The Android guidelines (as well as
the IDE) encourage us to *also* supplement it with "drawableLeft" for
compatibility with API <17.

However, when switching to RTL under API <23, this seems to be having the
effect of rendering *both* the "start" drawable, as well as the "left"
drawable... whereas theoretically the "start" drawable should override the
"left" drawable.

Anyway, it looks like it's perfectly acceptable to remove the "left"
drawable, and suppress the IDE check for it.  On API 17+ this means that
it will use the "start" drawable as expected, and on API <17 it still
renders the "start" drawable, but always places it on the left, which is
also expected.

Bug: T133169
Change-Id: Ia43adf10165543a9630c086c702005f1f07fc964