Update login and account creation to work with AuthManager
This updates the app's API calls to work with changes to the API
associated with the launch of AuthManager.

This does not remove any code but rather detects whether AuthManager is
enabled for the app site so that login and account creation can be handled
appropriately.  When legacy login and account creation are disabled on all
Wikipedias, the code to handle them can be removed accordingly.

Bug: T124229
Change-Id: Idcb8c8c05a8fa8321ce7bc67d3b4bea6393419f8