Fix optional table joins and unmarshalling
Several of the new table joins expect the rows of the right hand table
to be optional. This requires a _left_ join, not just a join.
Additionally, Cursors returned from join queries always have these
columns but they may be empty. Instead of checking for the presence of a
column name to determine whether the joined part exists or not, check
that a field with a nonnull constraint is nonnull.

Change-Id: I3da94a9953f880e03cac4b3ab40f9c8c73b21110