Fix Reading List page list membership queries
List memberships are preserved as a CSV:

  key,key,key

An empty lists mean delete and should never exist in the database and a
list of one has no commas:

  key

In practice, a list of three keys looks like:

  ABCDEF,ABCDEFjsAdifj,SDFpodskfposk

The prior implementation intended to perform a substring query on the
CSV where the query was wrapped in delimiters to restrict matches to
whole keys and prevent portions of keys matching. In the above example,
a match for the ABCDEF key should match the first item but not the
second, which only happens to contain ABCDEF as a substring. The query
looks like:

  %,ABCDEF,%

Since the first and last items of the key list won't be delimited on
both sides, the column contents should pre and postfixed with commas.
The prior implementation was confused and broken.

Change-Id: I4605e0e1d37aae873d171d637bea43467eec77b6