Hygiene: Use TextInputEditText instead of EditText.
This switches to using the slightly more proper TextInputEditText when
nesting inside a TextInputLayout. This allows us to place the "hint"
message in the parent TextInputLayout, instead of the EditText itself,
which works better when the IME is in "extract" mode.  We actually give
most of our EditTexts a flag of "flagNoExtractUi", but still, this
represents a hygienic improvement.

This also removes an annoying warning message in logcat about this issue.
No functional changes intended.

Bug: T133393
Change-Id: I57fe25fc95edbb906eb5cb186c01edbe81ce292f