Improve logic for storing and updating offline pages.
Previously, each page in a reading list would be marked as OUTDATED every
time it was upserted for any reason, which would cause unnecessary network
requests and disk writes.

This patch updates the logic to write the page to disk only once (when
it's added to a list), and then update the page only when the user visits
the page again while online.

Change-Id: I3d073a7788894e51e4f43f7a0cc64993beb85ae2