Add a small artificial delay after saving an edit.
After an edit is saved in the app, the change on the server side is not
immediate, and takes a few moments to start appearing in the response when
fetching the page content again.

Until this is fixed on the server side, we can work around it by forcing a
short delay right after a successful edit, before we attempt to read back
the updated page content.

Bug: T132803
Change-Id: I56758654aa5dd6890eb397de139140600c45bdbe