Make the message text in developer settings dialogs be highlightable.
When modifying developer settings, such as the RestBase URI format, we
show a message with a "hint" of what the setting should be. Since the hint
is relatively long, and contains many special characters, it's beneficial
to allow long-pressing of the text to highlight it and copy it into the
EditText field, and simply modify the pasted text.

Change-Id: Iaff11fe749553ac4640d709315a7f8ecf6a01762