Make reading lists sortable.
This adds a Toolbar button in the ReadingLists fragment used for sorting
the lists, as well as sorting the pages in a list. The button expands into
a menu with options to "sort by name" and "sort by recent".

If the user is looking at the list of lists, the sorting will apply to the
list of lists.  If the user is looking at the pages in a certain list, the
sorting will apply to the pages.

The sorting modes for both cases are stored in Prefs.

Bug: T136739
Change-Id: I76c095d912146fafbdd3f1e59de366767a00fafa