Fix possible crash by sharing cache between instances of OkHttpClient.
We create multiple instances of OkHttpClient, namely one per every Site
that the user visits.  And for each instance of this client, we create a
new Cache object, which points to the *same* cache directory. This
suggests that multiple caches can conflict with each other, and cause a
crash.

The OkHttp documentation[1] recommends using a single instance of
OkHttpClient, which we should eventually do, and it definitely recommends
using a single instance of Cache. And I quote:

"It is an error to have multiple caches accessing the same cache directory
simultaneously. Most applications should call new OkHttpClient() exactly
once, configure it with their cache, and use that same instance
everywhere."

[1] https://github.com/square/okhttp/wiki/Recipes#response-caching

Bug: T137077
Change-Id: Ia8bcc60fb019dbbb127da758dad9cf22996f8093