Check responses of synchronous network requests
A follow up to 9a21153, check the response statuses of synchronous
network requests. Probably simply because synchronous requests are less
common, I don't see checking the status called out explicitly in the
docs for the synchronous case. However, the status is checked in many
usages of the asynchronous Callback.onResponse(), and it seems
surprising not to for either case. There is also mention of checking the
synchronous status in this sample[0].

[0] https://github.com/square/retrofit/blob/0e4fe60/samples/src/main/java/com/example/retrofit/DeserializeErrorBody.java#L60

Change-Id: I37042a3d4a36b63988d350a90ff7d5df02bfaff4