"Because you read" feed card
This card displays up to three results from a morelike API search based
on a title from the user's recent browsing history (retrieved by a
subclass of our existing MainPageReadMoreTopicTask).

I implemented the search API call with Retrofit.  The existing "read
more" section still calls the API with an AsyncTask as before, but moving
all API searches to Retrofit could be a follow-up task.

I deferred on enforcing a minimum browsing time limit as specified in the
card for now, but in the future maybe this could be implemented by creating
a PageTimer and storing the page the user has most recently spent 30+
seconds on (or a queue of such pages) in a SharedPreference.

I also deferred on creating the "FeedManager" control logic since that will
likely be a substantial task in its own right.

Bug: T129080
Change-Id: I11ef1f59ee2530f116e7c02d2a238898210eb2ee