Fetch more feed contents automatically upon scroll.
This adds the piece of Feed interactivity where it automatically requests
more feed content when the user scrolls down sufficiently far.

Quick implementation note: I put the call to notifyDataSetChanged() into a
Runnable that is posted to the message queue. This is because the request
to fetch more feed contents is made during the view-binding phase of the
RecyclerView. Since some feed cards may return their contents immediately
(in the same thread), it may cause the RecyclerAdapter to be invalidated
while it's still laying out its contents. This will ensure that
notifyDataSetChanged() will get called only when it finishes its previous
layout.

Change-Id: I2667ad4aa4c59260ed117c4624aa413306b6d1ec