Remove 'Today' from nav menu
The feed is replacing 'Today' as the app's functional homescreen, and so
'Explore Wikipedia,' which links to it, is replacing it on the nav menu.

The appified main page is still accessible via the feed's main page card,
which is being added separately (no need to make it an explicit dependency
here, I think).

Bug: T129086
Change-Id: I75300d73951fb0f126f9c65307ca1841e235e9c2