Use Norwegian Bokmål language code not wiki code
A follow up to 8e9125b, the languages supported and namespace
translation scripts assume that the wiki subdomain matches the language
code. This is not true for Norwegian Bokmål which has a subdomain of
"no" and a language code of "nb". Update and rerun the scripts to record
the correct language code and move the "nb" to "no" language code to
subdomain translation from LanguageUtil to Site. A translation in
LanguageUtil is unwanted as "nb" is understood by the Android
configuration loader and only Site needs to be aware of the conversion.

Bug: T114042
Bug: T138720
Change-Id: Ia34c0791986e0b9904e178f599b43a4e06b503d6