Drop language name from card title strings
"Today on English Wikipedia", with English is passed in as a parameter, is problematic for translations
to languages which have different suffixes for genders/declinations.
When I switched to German language and refreshed the explore feed, it said
"Today on Deutsch Wikipedia". When translated it would say
"Heute auf Deutsch Wikipedia", but it should say
"Heute auf der deutschen Wikipedia" instead. (It also requires an article.)
Even then it sounds weird.
When I got to German Wikipedia it just says
"Willkommen bei Wikipedia" and not
"Willkommen bei der deutschen Wikipedia". The language name is not used.

I grant that if you switch languages often and for testing it would be nice to see which
WP the feed is based on. It feels awkward for most people, and would be grammatically wrong.

Change-Id: Id5d832d03cc9e7af62442cac1575ab9d23ae29ce