Featured image card: open image in gallery when clicked
Does what it says on the tin.  Further, logs gallery events with a constant
string ("FeedFeaturedImage") as the pageTitle value when opened in the
gallery.

Also includes some light refactoring to pass in the filename and Site as
String and Site rather than an "imageTitle" PageTitle object, which seems
confusing and unnecessary.

Bug: T132814
Change-Id: Ice34d54bf6c17f5359d707dddc148e24d1fdd4a0