Remove in-Wikitext image captions from news stories
The Wikitext we scrape for the In the News stories often includes a
parenthetical referral to a thumbnail for the section, which we don't use.
Instead, we get thumbnails for each card from the metadata for the linked
pages for each story.  This patch removes the inapplicable caption.

Note that this has a soft dependency on
I295c01b48a0d4feea6b21e07e85179deadf1f216 in that merging without it won't
hurt anything, but you'll need that patch applied to the service to see
the effect here.

Among current stories on enwiki, this applies to the one concerning the
Juno spacecraft.

Change-Id: I04284d257858f9d50494c8b0f4a459da099a2091