Update CardPageItem and consumers to reflect updated service output
In I5eb24688a774316481b98a494d3984044d09d948 we changed the output for
the random and featured article endpoints to include both the normalized
title in the normalizedtitle field, and the raw title in the title field
(where the normalized title previously was).  As an effect of this, right
now you'll see underscores in the Featured Article card article title.

This updates the relevant consumers to handle the updated output.

Change-Id: Id112b7ab904704ecb85d6cbe9b5f45e97fb2130f