Hygiene: infer nullity
There are seven public constructors to PageTitle, one of which accepts
raw JSON. Many paths of execution make it difficult to reason about a
class. Try to infer member variable nullity of PageTitle and add
annotations. For site, a bug existed where hashCode() expects it to be
nonnull but the JSON constructor sets it null. I've rewritten the
constructor to throw a RuntimeException in non-prod builds and use the
app site while remotely logging the error otherwise. I also changed a
couple of the hasProperties() conditionals to just check for properties
== null directly as the IDE isn't intelligent enough not to erroneously
flag subsequent dereferences.

Change-Id: Ibfff6290c6e5b7b9b35113d9688e15c76159fdbc