Remove hardcoded eye to nose offset on lead images
The upgrade to Fresco patch, 8287304, changed both the library and app
lead image implementations considerably with the side effect that some
of the assumptions around the finer details of face detection are
difficult to identify and verify. FaceDetection.getMidPoint() provides
the point between the subject's eyes. Previous to this patch, we
attempted to vertically offset to the nose as it gave a more pleasing
presentation on some articles. This seems to give a poor presentation
for other articles, so this patch removes it and also enables horizontal
offsets and corrects the default offsets from (0, 0) to (.5, .5). The
David Lloyd Johnston Japanese Wikipedia article gives a good example of
the horizontal offset in action.

For face detection to work properly, 1) the face must be detected
correctly and 2) that face must be favorably positioned relative the
bounds of the image. The image will not be zoomed in for the sake of
focus. For instance the face for the English Wikipedia governor-general
article is correctly identified but it's close to the picture edge so
the image cannot be offset sufficiently to be centered the face.

Bug: T139393
Change-Id: Ie11d8134a5443b2643a6e8a75a8e9bc4c013125e