Bugfix: Fix nearby crash on permissions request result
Moving the permission result listener to MainActivity in
I320c8ea6414d4edd7efcf3924505a89b93abae86 caused that listener to receive
result signals we weren't expecting, causing us to throw an exception when
hitting the default block.  Instead, let's not throw in the default block,
but allow signals not handled there to propagate to the fragments for
handling.

Bug: T140084
Change-Id: I39037f03755b787149360763b1332d354704d26b