Make continue reading Feed card images tappable
Continue reading Feed cards look almost identical to the first entry of
because you read cards but behaved differently. Continue reading cards
respond to taps on the image or the text but continue reading only
respond to text taps. Change CardLargeHeaderView.onClickListener() to
listen to the root View instead of the TextView and remove the cludge in
BecauseYouReadCardView.

Change-Id: If29e47fa671949f80c590454392aad5c45e465a5