Nail down face detection throughout the app.
This patch updates our custom face-detecting ImageView with a new default
handler that *automatically* centers the focus onto the face. All you need
to do is make sure that you set the image cropping to focusCrop in the
xml.

The patch updates the Featured Aricle card, Featured Image card, News
cards, and Continue Reading / Because You Read cards to all use automatic
face detection.

The actual face detection logic is now updated to properly center on the
nose of the subject, instead on the eyes, making the subjects appear
nicely centered in landscape-oriented views.

This also cleans up a lot of additional unnecessary logic surrounding the
previous incarnation of this feature.

Change-Id: Icf2385fd5de3b4f7b0a72dbb61dfabba2a5a1be6