Derive card dates from callback age param rather than static Date object
Whomever implemented the aggregated feed content client (spoiler: it was
me) unwisely chose to maintain a static Date object, set when a request was
made from the feed, that would then be also read from the callback to
assign dates to cards.  Unfortunately, this meant the date could be modified
before the callback returned.

This fixes that.

Bug: T140406
Change-Id: I4091945dc25b74bd79f5b59851f2ab39f6e87ce7