Hygiene: use default LinearLayout orientation
The default orientation of LinearLayouts are horizontal[0]. Defaults
should only be overridden when deviation is necessary. Consider the
recent case of overriding textColor on a TextView to black. What seemed
identical actually introduced a subtle bug apparent only in night mode.
Although not explicating code requires an understanding of the implicit
behavior, that is often a requirement regardless and less code is easier
to read and work with.

[0] https://developer.android.com/reference/android/widget/LinearLayout.html#attr_android:orientation

Change-Id: I3ad6a8937bc65759fdf514317e21a6633fe7b235