Fix memory leak related to HockeySDK
The current HockeySDK has a known issue that can leak an Activity, if the
MetricsManager is initialized with the Activity's context. For some
reason, MetricsManager asks for both an Application context and an
Activity context, when in fact it seems to function identically when
initialized with just the Application context.

https://github.com/bitstadium/HockeySDK-Android/issues/166

Bug: T139945
Change-Id: I059f1af67d5c5a6c1d2485eb5dd2d4f37cc42d1a