Convert remaining HttpURLConnection usages to OkHttp
Most of our core API calls already use the up-to-date Request/Response API
provided by OkHttp.  But there were a few remaining calls using the
HttpRequest library from earlier versions of the app.  This patch moves
more of them to Retrofit.

We won't be able to completely do away with OkUrlFactory until we stop
using the apps/android/java-mwapi library, which needs the cookie handling
that the OkHttpClient provides.  (In principle we could put together our
own implementation, but it's really not worth the trouble.)  Removing our
reliance on apps/android/java-mwapi will be in a patch to follow.

Also includes some light refactoring.

Bug: T134904
Change-Id: I785eeb329718453818d6d224fa5c059ce37cad2e