Try RB earlier after fallback
To further the RB roll out we want to try using RB earlier after a fallback case from RB to MW API has occurred.

This patch changes it so that the app tries to go back to RB after 5 successful page loads instead of waiting for a 100 successful page loads.

Change-Id: I3604f8e588c39c27cb1dd3967b75169ea7020b4f