Call MapView.onDestroy in Fragment.onDestroyView()
After switching to Butter Knife in 715b322, the MapView's lifecycle was
properly accounted for as Fragment.onCreateView() to onDestroyView(). In
8b63ffb, a null check was added to Fragment.onDestroy() where
MapView.onDestroy() was called per documentation. However, this can't be
correct because MapView will always be null at this time. Move the call
to Fragment.onDestroyView() instead of checking for null.

Change-Id: I509a57c3250e968af99ddd245f9356afac474fea