Fix possible infinite loop of saving page images.
When downloading images for a saved page, if any of the images fail to
download for any reason, that image url is removed from the urlMap (while
iterating through the map), which causes a ConcurrentModificationException.

Unfortunately, this condition also makes the download process start all
over again, and continue in an infinite loop.

This patch makes it no longer remove the failed url from the map, since
it's perfectly acceptable to leave it in (when the offline content is
loaded into the WebView, the broken image URL will simply not produce an
image, which is equivalent to the previous behavior).

Bug: T141892
Change-Id: I6061297c59e8db26354bec708b82a0890476e3f6