Finish decoupling Search fragment from MainActivity.
This completes the decoupling of the search-related fragments from
anything having to do with expecting a parent activity, and instead
using callback mechanisms.

The only thing that remains an indirect depenency on MainActivity is the
setSearchViewEnabled function, which expects a specific view hierarchy
inside the activity's toolbar. However, this will be removed and/or
simplified in subsequent patches as part of the navigation overhaul.

Change-Id: I07111a8dff5bbabaa0fa310d2d190c926c86ac7a