Change way to get title from links
For the mobileview/MW API case we can simply use the title
attribute instead of decoding the href attribute.

For the Parsoid/MCS case newer revisions of pages and once we get
all pages re-rendered after Icae4b83278791ed09284bb73c3ba80363cac9202
gets deployed will also have a title attribute.
For older pages the title attribute doesn't exists.
But we can derive it in JavaScript code from the href attribute by
removing a starting "/wiki/" or "./" and running decodeURIComponent on it.

That way the JS code always returns a title property, which the Java
side can just take as is.

The onUrlClick() method gets a new parameter titleString.
Instead of using getSite().titleForInternalLink(href) we now create the
PageTitle directly using titleString.

In the case of the gallery we don't get the link data from native code
instead of JS. The parsing of the HTML there is done using Html.fromHtml()
which doesn't preserve the title attribute of the link. In that one case
the Java code will derive the title from the href attribute like it did before.
This case is OK since the gallery data is not coming from Parsoid.

Introducing one factory method to create a PageTitle with a potential
fragment passed in separately. There was already a constructor with three
String parameters.

Related: Icae4b83278791ed09284bb73c3ba80363cac9202
Bug: T136223
Change-Id: Iedbe1ae611f4d433d4c6bcac3cce3935333753e0