Fix long-pressing of links inside the WebView.
I've noticed that long-pressing links was no longer working. This is
because we've recently made a change that updates the base URL of the
WebView dynamically, but then our long-press handler was expecting link
URIs to have an authority that exactly matches "wikipedia.org", which
would fail for language-specific domains, e.g. "en.wikipedia.org".

This patch fixes the URI checking logic, and thus fixes the long-press
handler.

Change-Id: I48b0add7bcaf21ad4e4db705d5823502205c2eb5