Create an activity with just PageFragment in it.
This will serve as the activity that is launched when browsing a page from
the new OverhaulActivity.

The new PageActivity is basically a copy of MainActivity, with unnecessary
stuff stripped out (nothing added).

This includes a very basic integration of the new activity with
OverhaulFragment, where tapping a link in the Feed will open the new
activity and load the page. Further integration will be in subsequent
patches, which will also fill in the remaining TODOs in PageActivity.

Bug: T141131
Change-Id: I76f99d25d359944c991d63d3a160df6bc1f50fd4