Restore lead image gradient (alternate version)
Previously, the lead image gradient was actually applied to the TextView
containing the article title/description, but that's now below the lead
image.  This adds a view on top of the lead image view with the desired
gradient.

If we want to adjust the opacity now that the text is moved, we can do
that separately.

Bug: T135294
Change-Id: I5cbb7e679b4efe23672fbc9351d658ef2ab45441