Add screenshot unit test support and diff script
Add support for screenshot unit tests to detect regressions and a diff
script for identifying them.

The Facebook plugin, although image diffs are shown in the
presentation[0], does not in itself provide them. However, the plugin is
still wanted for its on device Java API. The screenshot diff script
supplements.

Some considerations were made in writing it. An ideal place for image
diffing support would be in the plugin itself, perhaps using Pillow's
ImageChops since it's already a dependency[1,2]. This effort was not
pursued as the plugin bails on the first error and the expected effort
seemed large.

The diff script uses the popular ImageMagick convert command and is
written in Make to test only changed files and easily invoke convert.
A more ideal and lasting solution can be made when tests are being made
regularly.

This is not yet wired into CI but shouldn't cause unwanted errors
either.

[0] http://facebook.github.io/screenshot-tests-for-android/#getting-started
[1] https://github.com/facebook/screenshot-tests-for-android/blob/v0.4.1/plugin/src/py/android_screenshot_tests/recorder.py#L82
[2] https://github.com/facebook/screenshot-tests-for-android/blob/master/README.md#requirements

Bug: T118102
Change-Id: Ib9e25ff0a8d0c1fe7dc61836bf2dd07b71e4f223