Rip out MainActivity, et al.
This removes the old MainActivity, and delegates any of its remaining
references to either PageActivity or OverhaulActivity.

This also rips out other things from the pre-overhaul navigation model,
including the old NavDrawerHelper, RandomHandler, Search fragment, toolbar
hiding handler, etc.

In the next patch, we'll rename OverhaulActivity to MainActivity and
remove any other "overhaul" designations.

And in the patch after that, we'll start solidifying which types of
intents will be sent to which activity.

Bug: T143441
Change-Id: Ibb6003b6bcca2fe4298df3c8e2e7e578fa2de2d8