Hook up hiding and fading of Page toolbar.
I tried using a CoordinatorLayout to control the behavior of the Toolbar
with respect to the WebView; I really did! But my best attemps proved
futile. The most reasonable solution for now is simply to use the old
"hide-handler" pattern, so I brought back the old HideHandler and hooked
it up to the Toolbar in PageActivity.

This also adds a bit of functionality: when you're scrolled deeply into a
page (with the toolbar hidden), you can now click anywhere within the
WebView, and it will cause the toolbar to appear again (scrolling is no
longer necessary to show it).

Change-Id: Ia0a427c58790fb8e7bf748a1c12963c9b7788604