Use a proper ActionMode for searching History and Reading Lists.
Our loyal users who have 1000+ pages in reading lists can now search them!

This adds a "search" button to the actual Toolbar in the History and
Reading Lists fragments, which takes the user to a proper ActionMode that
dispatches search queries.

Searching of reading lists is two-pronged, since it depends on whether the
user is looking at the list of lists or the pages in a particular list.
When looking at the list of lists, I was able to tweak the db query that
fetches the list so that it includes the search term. Unfortunately the
same cannot be done when looking at individual list pages; in this case,
we're manually searching the pages and showing the results.

Bug: T145100
Change-Id: Iebf78a0bda3f92774ada0faac479053e105dcc99