Hygiene: Decouple user mgmt from app and storage
Having the user management functionality tied to the Application instance
and the storage in shared preferences makes it really hard to create mock
user for unit testing.

While the User class is now a bit more complex the resulting usages
are now much simpler and concerns are clearer.

Added a simple unit test for user management.

This patch is in preparation for unit testing the work for T132834.

Bug: T132834
Change-Id: I1c45843b94e0bbdc5205f9cd20677ae942303e70