Retrofit LoginTask
It seems easier to chain request implemented using Retrofit than the traditional
java-mwapi requests. In the future we want to chain yet another request, in
addition to the two we have already. One to get the login token.
The 2nd one for the actual login request.

This patch replaces the LoginTask with a Retrofit2 LoginClient.
The test changes from LoginTaskTest to LoginClientTest.

Bug: T132834
Change-Id: I6d9818c92bfd834adf966a6c5808b7c0a236235d