Hygiene: replace ImageButton with ImageView
Replace ImageButtons with ImageViews and add the appropriate styling
where needed. ImageButton is nearly identical to ImageView. It seems
more appropriate to use ImageViews with styles than subclasses for such
little gain. Also remove some unused ImageView attributes, replace a
AppCompatImageView declaration with ImageView which is the implicit
implementation, and replace a usage of #ffffff with the platform white
symbol.

Change-Id: Ia79c02c211e464d2c8310bd7961dda71b710d237