Remove random from aggregated feed parsing
We had random data display in earlier versions of the app but don't actually use it.
During an RB deploy today where we wanted to change the format to an array
we discovered that the app was still expecting the random property to be an object.

Therefore let's remove the random property from even being read in
to allow for more flexibility in the future.

Change-Id: I56a33249d0083c88eec3254aca9173a7c8edfbd6