Give a text label to the Feed overflow menu button.
Because our "custom" overflow button uses @null as its title, it can lead
to some unexpected behavior. For example, in the rare case when this
button is forced to overflow into the "real" overflow menu provided by
Android, the resulting menu item appears empty.

This patch simply provides a text label for the custom overflow button, so
that if it overflows "for real", it will at least have a label.

Bug: T145911
Change-Id: If5bf1c3b4918dd29af1ebd146b156bc704d0732b