Allow Dark mode for MainActivity.
There was really only one color attribute that needed to be introduced to
support dark mode.

This also requires us to set the statusBarColor programmatically, since
the theme is now set dynamically (and it's shared with other activities
that might not want a custom statusBarColor). Might be the subject of a
future refactor.

Bug: T145686
Bug: T146109
Change-Id: Ia952336b043aec0be39743cb10e2f0fb3a574b77