Don't require faketouch support
Previously, we required faketouch support to allow devices that don't
have a touchscreen but do offer mouse support. However, faketouch to
certain Amazon devices seems to be a requirement in *addition* to
optional touchscreen support. The console warning was:

  This APK’s manifest conflicts with 3 of your targeted Amazon Fire
  devices.
  Starting in October, any APK update will require an updated manifest
  to maintain your current device support. Click on “View device
  support” below for more information on how to address conflicts.

The warning detail is that we require devices to have
android.hardware.faketouch but these devices don't have it:

  Kindle Fire (2011)
  Kindle Fire HD 8.9 (2012)
  Kindle Fire HD 8.9 WAN (2012)
  And a couple non-Amazon devices

Change-Id: I94753fddac75b866219834f993d723a3224a698b